Implemented OpenCL brute force string search
So, it's not yet KMP because I wanted to get OpenCL basics out of the
way first.

It's heavily modified snippet from
[here](https://rocmdocs.amd.com/en/latest/Programming_Guides/Opencl-programming-guide.html#example-saxpy-function).
There is some dead code still, I'll merge it once I clean it up, but it
works and can be run, and hopefully read and understood. I'll have a KMP
proper done later this week.