Cleaned up Main.cpp and changed path to cl kernel
Path was changed because when running in CMake, original file in source
dir was inaccessible. To keep things simple, file was copied, but
`$CMAKE_CURRENT_BINARY_DIR` doesn't refer to actual executable dir, but
to a build dir, so the path looks like this:

```
- string_search_kernel.cl
- Debug
+ --- HelloOpenCL.exe
```

So path had to be changed to accomodate this.